Get rid of the dialog on completion of unsubscribing and removed unsubscribed repos from view

+++ app/controllers/api.go
@@ -20,30 +20,34 @@ func (c Api) GetAuthenticatedUser() *models.User {
 type ApiResult struct {
 	Success bool   `json:"success"`
 	Error   string `json:"error,omitempty"`
+	Owner   string `json:"owner"`
+	Repo    string `json:"repo"`
 }
 
-func NewApiResultSuccess() ApiResult {
-	return ApiResult{true, ""}
+func NewApiResultSuccess(owner string, repo string) ApiResult {
+	return ApiResult{true, "", owner, repo}
 }
 
-func (c Api) Unsubscribe(owner string, repo string) revel.Result {
+func (c Api) Unsubscribe(owner string, repo string, dryRun bool) revel.Result {
 	u := c.GetAuthenticatedUser()
 	if u == nil {
 		c.Response.Status = 401
-		return c.RenderJson(ApiResult{false, "Not authenticated"})
+		return c.RenderJson(ApiResult{false, "Not authenticated", owner, repo})
 	}
 
 	client := GithubClientForUser(u)
 
 	if repo == "" || owner == "" {
-		return c.RenderJson(ApiResult{false, "Repo and repo owner is required"})
+		return c.RenderJson(ApiResult{false, "Repo and repo owner is required", "owner", "repo"})
 	}
 
-	_, err := client.Activity.DeleteRepositorySubscription(owner, repo)
-	if err != nil {
-		revel.ERROR.Println(err)
-		return c.RenderJson(ApiResult{false, err.Error()})
+	if !dryRun {
+		_, err := client.Activity.DeleteRepositorySubscription(owner, repo)
+		if err != nil {
+			revel.ERROR.Println(err)
+			return c.RenderJson(ApiResult{false, err.Error(), owner, repo})
+		}
 	}
 
-	return c.RenderJson(NewApiResultSuccess())
+	return c.RenderJson(NewApiResultSuccess(owner, repo))
 }
